--- jaffle_shop/.dbt/profiles.yml
@@ -1,12 +1,0 @@
-jaffle_shop: # this needs to match the profile in your dbt_project.yml file
-    target: dev
-    outputs:
-        dev:
-            type: bigquery
-            method: oauth
-            project: preqldata # Replace this with your project id
-            dataset: dbt_test # Replace this with dbt_your_name, e.g. dbt_bilbo
-            threads: 4
-            timeout_seconds: 300
-            location: US
-            priority: interactive
