--- jaffle_shop/.gitignore
@@ -1,4 +1,0 @@
-
-target/
-dbt_packages/
-logs/
