--- jaffle_shop/models/customer/schema.yml
@@ -1,23 +1,0 @@
-models:
-- columns:
-  - description: No description provided
-    name: customer_id
-    tests: []
-  - description: No description provided
-    name: customer_first_name
-    tests: []
-  - description: No description provided
-    name: customer_last_name
-    tests: []
-  - description: No description provided
-    name: customer_first_order_date
-    tests: []
-  - description: No description provided
-    name: customer_most_recent_order_date
-    tests: []
-  - description: No description provided
-    name: customer_number_of_orders
-    tests: []
-  description: Automatically generated model from preql
-  name: customers_preql_gen_model
-version: 2
