--- jaffle_shop/models/example/customer.preql
@@ -1,29 +1,0 @@
-import order as orders;
-
-key id int;
-property id.first_name string;
-property id.last_name string;
-
-property id.first_order_date <-min(orders.date) by id;
-property id.most_recent_order_date <- max(orders.date) by id;
-property id.raw_number_of_orders <- count(orders.id) by id;
-
-datasource customers_raw (
-    id:id,
-    first_name:first_name,
-    last_name:last_name
-)
-grain (id)
-address dbt-tutorial.jaffle_shop.customers
-;
-
-
-persist customers into preqldata.dbt_test.customers_preql from 
-SELECT
-        id,
-        first_name,
-        last_name,
-        first_order_date,
-        most_recent_order_date,
-        coalesce(raw_number_of_orders, 0) -> number_of_orders
-;
