--- preqlt/__init__.py
@@ -1,7 +1,0 @@
-from preqlt.core import enrich_environment
-from preqlt.dbt.generate_dbt import generate_model as generate_dbt_model
-from preqlt.dbt.config import DBTConfig
-
-__version__ = "0.0.3"
-
-__all__ = ["enrich_environment", "generate_dbt_model", "DBTConfig"]
