+++ trilogyt/__init__.py
@@ -1,0 +1,7 @@
+from trilogyt.core import enrich_environment
+from trilogyt.dbt.generate_dbt import generate_model as generate_dbt_model
+from trilogyt.dbt.config import DBTConfig
+
+__version__ = "0.0.3"
+
+__all__ = ["enrich_environment", "generate_dbt_model", "DBTConfig"]
