+++ trilogyt/constants.py
@@ -1,7 +1,7 @@
 from logging import getLogger
 from logging import StreamHandler, INFO
 
-PREQLT_NAMESPACE = "_preqlt"
+TRILOGY_NAMESPACE = "_trilogyt"
 
 OPTIMIZATION_NAMESPACE = "optimization"
