+++ trilogyt/core.py
@@ -1,13 +1,13 @@
 from trilogy.core.models import Environment, Concept, DataType
 from trilogy.core.enums import Purpose
-from preqlt.enums import PreqltMetrics
+from trilogyt.enums import PreqltMetrics
 from trilogy.core.functions import CurrentDatetime
-from preqlt.constants import PREQLT_NAMESPACE
+from trilogyt.constants import TRILOGY_NAMESPACE
 
 ENVIRONMENT_CONCEPTS = [
     Concept(
         name=PreqltMetrics.CREATED_AT.value,
-        namespace=PREQLT_NAMESPACE,
+        namespace=TRILOGY_NAMESPACE,
         datatype=DataType.DATETIME,
         purpose=Purpose.CONSTANT,
         lineage=CurrentDatetime([]),
@@ -16,8 +16,8 @@ ENVIRONMENT_CONCEPTS = [
 
 
 def enrich_environment(env: Environment):
-    if PREQLT_NAMESPACE in env.imports:
-        raise Exception(f"{PREQLT_NAMESPACE} namespace already exists")
+    if TRILOGY_NAMESPACE in env.imports:
+        raise Exception(f"{TRILOGY_NAMESPACE} namespace already exists")
 
     for concept in ENVIRONMENT_CONCEPTS:
         env.add_concept(concept)
